+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/aggregate/Aggregation.java
@@ -15,26 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.connector.expressions;
+package org.apache.spark.sql.connector.expressions.aggregate;
+
+import java.io.Serializable;
 
 import org.apache.spark.annotation.Evolving;
+import org.apache.spark.sql.connector.expressions.FieldReference;
 
 /**
- * An aggregate function that returns the minimum value in a group.
+ * Aggregation in SQL statement.
  *
  * @since 3.2.0
  */
 @Evolving
-public final class Min implements AggregateFunc {
-  private final FieldReference column;
+public final class Aggregation implements Serializable {
+  private final AggregateFunc[] aggregateExpressions;
+  private final FieldReference[] groupByColumns;
 
-  public Min(FieldReference column) { this.column = column; }
+  public Aggregation(AggregateFunc[] aggregateExpressions, FieldReference[] groupByColumns) {
+    this.aggregateExpressions = aggregateExpressions;
+    this.groupByColumns = groupByColumns;
+  }
 
-  public FieldReference column() { return column; }
+  public AggregateFunc[] aggregateExpressions() { return aggregateExpressions; }
 
-  @Override
-  public String toString() { return "MIN(" + column.describe() + ")"; }
-
-  @Override
-  public String describe() { return this.toString(); }
+  public FieldReference[] groupByColumns() { return groupByColumns; }
 }
