+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/aggregate/CountStar.java
@@ -15,36 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.connector.expressions;
+package org.apache.spark.sql.connector.expressions.aggregate;
 
 import org.apache.spark.annotation.Evolving;
 
 /**
- * An aggregate function that returns the number of the specific row in a group.
+ * An aggregate function that returns the number of rows in a group.
  *
  * @since 3.2.0
  */
 @Evolving
-public final class Count implements AggregateFunc {
-  private final FieldReference column;
-  private final boolean isDistinct;
+public final class CountStar implements AggregateFunc {
 
-  public Count(FieldReference column, boolean isDistinct) {
-    this.column = column;
-    this.isDistinct = isDistinct;
+  public CountStar() {
   }
-
-  public FieldReference column() { return column; }
-  public boolean isDistinct() { return isDistinct; }
 
   @Override
-  public String toString() {
-    if (isDistinct) {
-      return "COUNT(DISTINCT " + column.describe() + ")";
-    } else {
-      return "COUNT(" + column.describe() + ")";
-    }
-  }
+  public String toString() { return "COUNT(*)"; }
 
   @Override
   public String describe() { return this.toString(); }
